GasPvtMultiplexer: move implementation to impl file
